% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc_cal}
\alias{auc_cal}
\title{Calculate AUC using Linear-Log Trapezoidal Method}
\usage{
auc_cal(data, startpoint, endpoint, auc.type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{startpoint}{Numeric. First timepoint.}

\item{endpoint}{Numeric. Second timepoint.}

\item{auc.type}{\code{auc} or \code{aumc}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC/AUMC
}
\description{
Calculate AUC using Linear-Log Trapezoidal Method
}
\keyword{internal}
