% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaZ.R
\name{calculateRsqAdj}
\alias{calculateRsqAdj}
\title{Calculate the parameter reports the goodness of fit statistic for the terminal elimination phase, adjusted for the number of points used in the estimation of Lambda Z}
\usage{
calculateRsqAdj(X, Y)
}
\arguments{
\item{X}{Numeric vector of time points at the terminal elimination phase}

\item{Y}{Numeric vector of concentrations at the terminal elimination phase}
}
\value{
A single numeric value representing the goodness of fit statistic for the terminal elimination phase
}
\description{
Calculate the parameter reports the goodness of fit statistic for the terminal elimination phase, adjusted for the number of points used in the estimation of Lambda Z
}
\keyword{internal}
