% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{dataQc}
\alias{dataQc}
\title{Data checking and pre-treatment
Sorting: prior to analysis, data within each profile is sorted in ascending time order
Inserting initial time points
Data exclusions: 1. Missing values, 2. Data points preceding the dose time}
\usage{
dataQc(time, conc, sampletype)
}
\arguments{
\item{time}{Numeric vector of time points}

\item{conc}{Numeric vector of concentrations}

\item{sampletype}{sample type: "tcell" or "cytokine"}
}
\value{
QCed data.table
}
\description{
Data checking and pre-treatment
Sorting: prior to analysis, data within each profile is sorted in ascending time order
Inserting initial time points
Data exclusions: 1. Missing values, 2. Data points preceding the dose time
}
\keyword{internal}
