% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAuc12M}
\alias{getAuc12M}
\title{Calculate AUC from day0 to month12}
\usage{
getAuc12M(data, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUC from day0 to month12
}
\keyword{internal}
