% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAuc6M}
\alias{getAuc6M}
\title{Calculate AUC from day0 to month6}
\usage{
getAuc6M(data, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUC from day0 to month6
}
\keyword{internal}
