% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAucAll}
\alias{getAucAll}
\title{Calculate AUC from the time of dosing to the time of the last observation. If the last concentration is positive, AUClast=AUCall.}
\usage{
getAucAll(data, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUC from the time of dosing to the time of the last observation. If the last concentration is positive, AUClast=AUCall.
}
\keyword{internal}
