% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAucInf}
\alias{getAucInf}
\title{Calculate AUC from time of dosing extrapolated to infinity, based on the last observed concentration (_obs) or last predicted concentration (_pred). AUClast + (Clast/Lambda_z)}
\usage{
getAucInf(data, type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUC from time of dosing extrapolated to infinity, based on the last observed concentration (_obs) or last predicted concentration (_pred). AUClast + (Clast/Lambda_z)
}
\keyword{internal}
