% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAucInfD}
\alias{getAucInfD}
\title{Calculate AUCINF/Dose}
\usage{
getAucInfD(data, type, Dose, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{Dose}{dosage}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUCINF/Dose
}
\description{
Calculate AUCINF/Dose
}
\keyword{internal}
