% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAucLastD}
\alias{getAucLastD}
\title{Calculate AUClast/Dose}
\usage{
getAucLastD(data, Dose, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{Dose}{Dosage}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUClast/Dose
}
\keyword{internal}
