% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAucPrecentExtrap}
\alias{getAucPrecentExtrap}
\title{Calculate percentage of AUCINF(_obs, _pred) due to extrapolation from Tlast to infinity, 100\link{(AUCINF– AUClast)/AUCINF}}
\usage{
getAucPrecentExtrap(data, type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated percentage of AUCINF(_obs, _pred) due to extrapolation from Tlast to infinity
}
\description{
Calculate percentage of AUCINF(_obs, _pred) due to extrapolation from Tlast to infinity, 100\link{(AUCINF– AUClast)/AUCINF}
}
\keyword{internal}
