% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAumcInf}
\alias{getAumcInf}
\title{Calculate Area under the first moment curve (AUMC) extrapolated to infinity, based on the last observed concentration (obs) or the last predicted concentration (pred)}
\usage{
getAumcInf(data, type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUMC
}
\description{
Calculate Area under the first moment curve (AUMC) extrapolated to infinity, based on the last observed concentration (obs) or the last predicted concentration (pred)
}
\keyword{internal}
