% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAumcLast}
\alias{getAumcLast}
\title{Calculate area under the moment curve from the time of dosing to the last measurable (positive) concentration}
\usage{
getAumcLast(data, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated AUMC
}
\description{
Calculate area under the moment curve from the time of dosing to the last measurable (positive) concentration
}
\keyword{internal}
