% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getAumcPrecentExtrap}
\alias{getAumcPrecentExtrap}
\title{Calculate Percent of AUMCINF(_obs, _pred) that is extrapolated}
\usage{
getAumcPrecentExtrap(data, type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing Percent of AUMCINF(_obs, _pred) that is extrapolated
}
\description{
Calculate Percent of AUMCINF(_obs, _pred) that is extrapolated
}
\keyword{internal}
