% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getClast}
\alias{getClast}
\title{Extract the parameter reports the observed concentration corresponding to Tlast}
\usage{
getClast(data)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}
}
\value{
A single numeric value representing the observed concentration corresponding to Tlast
}
\description{
Extract the parameter reports the observed concentration corresponding to Tlast
}
\keyword{internal}
