% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getClF}
\alias{getClF}
\title{Calculate total body clearance for extravascular administration, Dose/AUCINF}
\usage{
getClF(data, type, Dose, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{Dose}{dosage}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calcuated clearance
}
\description{
Calculate total body clearance for extravascular administration, Dose/AUCINF
}
\keyword{internal}
