% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getCmax}
\alias{getCmax}
\title{Extract the parameter reports the maximum observed concentration}
\usage{
getCmax(data)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}
}
\value{
A single numeric value representing the maximum observed concentration
}
\description{
Extract the parameter reports the maximum observed concentration
}
\keyword{internal}
