% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getCmaxD}
\alias{getCmaxD}
\title{Calculate the parameter reports the ratio of Cmax and dosage}
\usage{
getCmaxD(data, Dose)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{Dose}{Dosage}
}
\value{
A single numeric value representing Cmax/Dose
}
\description{
Calculate the parameter reports the ratio of Cmax and dosage
}
\keyword{internal}
