% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getMrtInf}
\alias{getMrtInf}
\title{Calculate mean residence time (MRT) extrapolated to infinity based on AUCINF(_obs, _pred)}
\usage{
getMrtInf(data, type, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calcuated MRT
}
\description{
Calculate mean residence time (MRT) extrapolated to infinity based on AUCINF(_obs, _pred)
}
\keyword{internal}
