% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getMrtLast}
\alias{getMrtLast}
\title{Calculate mean residence time from the time of dosing to the time of the last measurable concentration. AUC_0_last / AUC_infinity}
\usage{
getMrtLast(data, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calculated MRT
}
\description{
Calculate mean residence time from the time of dosing to the time of the last measurable concentration. AUC_0_last / AUC_infinity
}
\keyword{internal}
