% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getSampleNumber}
\alias{getSampleNumber}
\title{Extract the parameter reports the number of non-missing observations used in the analysis of the profile}
\usage{
getSampleNumber(data)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}
}
\value{
A single numeric value representing the number of observations
}
\description{
Extract the parameter reports the number of non-missing observations used in the analysis of the profile
}
\keyword{internal}
