% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getTlast}
\alias{getTlast}
\title{Extract the parameter reports the time of last measurable (positive) observed concentration}
\usage{
getTlast(data)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}
}
\value{
A single numeric value representing the time of last measurable (positive) observed concentration
}
\description{
Extract the parameter reports the time of last measurable (positive) observed concentration
}
\keyword{internal}
