% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalParameters.R
\name{getTmax}
\alias{getTmax}
\title{Extract the parameter reports the time of maximum observed concentration}
\usage{
getTmax(data)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}
}
\value{
A single numeric value representing the time of maximum observed concentration
}
\description{
Extract the parameter reports the time of maximum observed concentration
}
\keyword{internal}
