% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{getVzF}
\alias{getVzF}
\title{Calculate volume of distribution based on the terminal phase, Dose / (AucInf * LambdaZ)}
\usage{
getVzF(data, type, Dose, method)
}
\arguments{
\item{data}{A \code{data.table} with columns \code{time} and \code{conc}.
Time must be numeric and strictly increasing.}

\item{type}{\code{pred} or \code{obs}}

\item{Dose}{dosage}

\item{method}{\code{LinearUpLogDown}, \code{Linear}, \code{Log}}
}
\value{
A single numeric value representing calcuated volume of distribution
}
\description{
Calculate volume of distribution based on the terminal phase, Dose / (AucInf * LambdaZ)
}
\keyword{internal}
