% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{linearauc}
\alias{linearauc}
\title{Calculate AUC using Linear Trapezoidal Method}
\usage{
linearauc(c1, c2, t1, t2)
}
\arguments{
\item{c1}{Numeric. Concentration at time \code{t1}.}

\item{c2}{Numeric. Concentration at time \code{t2}.}

\item{t1}{Numeric. First timepoint.}

\item{t2}{Numeric. Second timepoint.}
}
\value{
A single numeric value representing calculated AUC
}
\description{
Calculate AUC using Linear Trapezoidal Method
}
\keyword{internal}
