% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{linearaumc}
\alias{linearaumc}
\title{Calculate AUMC (Area under the first moment curve) using Linear Trapezoidal Method}
\usage{
linearaumc(c1, c2, t1, t2)
}
\arguments{
\item{c1}{Numeric. Concentration at time \code{t1}.}

\item{c2}{Numeric. Concentration at time \code{t2}.}

\item{t1}{Numeric. First timepoint.}

\item{t2}{Numeric. Second timepoint.}
}
\value{
A single numeric value representing calculated AUMC
}
\description{
Calculate AUMC (Area under the first moment curve) using Linear Trapezoidal Method
}
\keyword{internal}
