% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{logestimate}
\alias{logestimate}
\title{Log-linear concentration interpolation (internal)}
\usage{
logestimate(c1, c2, t1, t2, timepoint)
}
\arguments{
\item{c1}{Numeric. Concentration at time \code{t1}.}

\item{c2}{Numeric. Concentration at time \code{t2}.}

\item{t1}{Numeric. First timepoint.}

\item{t2}{Numeric. Second timepoint.}

\item{timepoint}{Numeric. The timepoint at which to estimate the concentration.}
}
\value{
Estimated concentration at the specified \code{timepoint}.
}
\description{
Estimates a concentration at a given timepoint using log-linear interpolation
between two known concentrations and their corresponding timepoints.
}
\keyword{internal}
